+++ NEWS.md
@@ -1,6 +1,11 @@
 # ggpedigree (dev version)
+
+# ggpedigree 0.8.0.1
 * Updated JOSS paper
-* Refactored step 2 of the ggpedigree function to .transformPed()
+* Refactored step 2+3 of the ggpedigree function to .transformPed()
+* Renamed p variable as plotObject in ggpedigree subfunctions
+* Smarter implementation of aliases
+* Updated hex to actually use ggpedigree
 
 # ggpedigree 0.8.0
 * Makes ggrepel a suggested package instead of a required one.